pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stability-ai"
version = "0.0.1"
description = "Python client for Stability AI API"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Jack Beoris" }]
requires-python = ">=3.8"
dependencies = [
    "python-dotenv~=1.0",
    "requests~=2.32"
]
optional-dependencies = { dev = [
    "pytest"
] }

[project.urls]
homepage = "https://jackbeoris.com"
repository = "https://github.com/jbeoris/stability-ai-python-sdk"

[tool.setuptools]
packages = ["stability_ai"]

[tool.pytest.ini_options]
testpaths = "tests/"